+++ duilib/CEFControl/CefWindowUtils_MacOS.cpp
@@ -1,0 +1,46 @@
+#include "CefWindowUtils.h"
+#include "duilib/Core/Window.h"
+
+#ifdef DUILIB_BUILD_FOR_MACOS
+
+namespace ui
+{
+void SetCefWindowPos(CefWindowHandle cefWindow, CefControl* pCefControl)
+{
+    if ((cefWindow == 0) || (pCefControl == nullptr)) {
+        return;
+    }
+    Window* pWindow = pCefControl->GetWindow();
+    if (pWindow == nullptr) {
+        return;
+    }
+    
+}
+
+void SetCefWindowVisible(CefWindowHandle cefWindow, CefControl* pCefControl)
+{
+    if ((cefWindow == 0) || (pCefControl == nullptr)) {
+        return;
+    }
+    Window* pWindow = pCefControl->GetWindow();
+    if (pWindow == nullptr) {
+        return;
+    }
+    
+}
+
+void SetCefWindowParent(CefWindowHandle cefWindow, CefControl* pCefControl)
+{
+    if ((cefWindow == 0) || (pCefControl == nullptr)) {
+        return;
+    }
+    Window* pWindow = pCefControl->GetWindow();
+    if (pWindow == nullptr) {
+        return;
+    }
+    
+}
+
+} //namespace ui
+
+#endif //DUILIB_BUILD_FOR_MACOS
