+++ duilib/duilib.vcxproj.filters
@@ -684,6 +684,9 @@
     <ClCompile Include="CEFControl\CefManager_Windows.cpp">
       <Filter>CEFControl\Windows</Filter>
     </ClCompile>
+    <ClCompile Include="CEFControl\CefWindowUtils_Windows.cpp">
+      <Filter>CEFControl\Windows</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Animation\AnimationManager.h">
@@ -1426,6 +1429,9 @@
     </ClInclude>
     <ClInclude Include="CEFControl\CefManager_Windows.h">
       <Filter>CEFControl\Windows</Filter>
+    </ClInclude>
+    <ClInclude Include="CEFControl\CefWindowUtils.h">
+      <Filter>CEFControl</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
